feat: Cache source infos for faster build graph calculation
Each build graph calculation requires us to read a specific commit from all ~12k source repos, and to parse the source info file in that commit.
This work is almost completely duplicated between different namespaces, as almost all of the nodes in the global dependency graph are based on the source info file from the respective main branch.

This change improves the situation by implementing a cache for `SourceInfo` structs and commit hashes associated with each branch. To evade potential cache invalidation bugs, the cache is extremely short-lived: each time we calculate a new build graph for every namespace, the cache is recreated from scratch. This means that the first build graph to be computed in each batch takes as long as before, but all subsequent graph calculations in that batch will be significantly faster.
The main improvement here is that the server will be able to handle large numbers of build namespaces better.