fix(graph): visit splitpkg node edges even if pkgbase has been visited
Previously we were skipping on all edges from the split packages as we
were only visiting the edges of the current node, if the graph did not
already contain a node for the current packages. This doesn't work, as
the nodes in this build graph are pkgbase nodes, subsequently we were
missing out on all the edges from all other split packages except the
one whose name coincidently matches with the pkgbase itself.

Visit all the current node edges regardless if the pkgbase node is
already present in the graph or not.

Fixes #121