args: add option to control tokio console telemetry
Turn off tokio console telemetry be default, as due to its nature its
hogging quite some memory for the historical async task collection.
This could be fine tuned for the particular application use cases, as
the defaults store quite a lot of data up to like 1h, but for simplicity
we simply disable it by default for now and retain an easy switch for
investigation.