fix(repo-updated): fix certain repos not being updated due to sorting
The GraphQL query is sorting by last_activity_at, but we are iterating
sequentially through updated_at while assuming its in perfect order to
be able to early exit.
The sorting of a different field lead to entries being put in wrong
order which lead to an early exit and some repos being skipped.
Unfortunately the GraphQL API does not allow to sort namespace projects
by updated_at, hence we need to migrate to last_activity_at which has
the side effect its only persistent max once an hour plus contains way
more events compared to the updated_at. Subsequently we will pull more
repos then strictly needed.

As a followup, we should contribute a path or issue to GitLab to be able
to sort by updated_at.

Fixes #180